# RestClientSDKLibraryHW5/Swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "HW5 Image Porducer API",
    "description": "Image Produvcer for Assignment V",
    "version": "v1"
  },
  "paths": {
    "/api/v1/uploadedimages/{imageConversionMode}": {
      "put": {
        "tags": [
          "ImageProducer"
        ],
        "summary": "Uploads a file",
        "operationId": "UploadFile",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageConversionMode",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "fileData",
            "type": "file"
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ImageProducer.DataTransferObjects.ErrorResponse"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ImageProducer.DataTransferObjects.ErrorResponse": {
      "type": "object",
      "properties": {
        "errorNumber": {
          "format": "int32",
          "type": "integer"
        },
        "parameterName": {
          "type": "string"
        },
        "parameterValue": {
          "type": "string"
        },
        "errorDescription": {
          "type": "string"
        }
      }
    }
  }
}